// src/containers/types/Container.ts
/**
 * The contents of this file are subject to the CYPHON Proprietary Non-
 * Commercial Registered User Use License Agreement (the "Agreement”). You
 * may not use this file except in compliance with the Agreement, a copy
 * of which may be found at https://github.com/dunbarcyber/cyclops/. The
 * developer of the CYPHON technology and platform is Dunbar Security
 * Systems, Inc.
 *
 * The CYPHON technology or platform are distributed under the Agreement on
 * an “AS IS” basis, WITHOUT WARRANTY OF ANY KIND, either express or
 * implied. See the Agreement for specific terms.
 *
 * Copyright (C) 2017 Dunbar Security Solutions, Inc. All Rights Reserved.
 *
 * Contributor/Change Made By: ________________. [Only apply if changes
 * are made]
 */

// Local
import { Field } from '../../cyphon/types/Response';
import { NormalizedList } from '../../common/types/normalizr';
import { Dictionary } from '../../common/types/object';

export interface Container {
  id: number;
  name: string;
  url: string;
  bottle: number;
  label: number;
  taste: number;
  fields: Field[] | string[];
}

export interface ContainerNested extends Container {
  fields: Field[];
}

export interface ContainerFlat extends Container {
  fields: string[];
}

export interface NormalizedContainerEntities {
  containers: Dictionary<ContainerFlat>;
  fields: Dictionary<Field>;
}

export type NormalizedContainerList = NormalizedList<number, NormalizedContainerEntities>;